+++ src/repo/stock.ts
@@ -4,11 +4,18 @@ import * as factory from '@motionpicture/ttts-factory';
 import StockModel from '../repo/mongoose/model/stock';
 
 /**
- * 在庫レポジトリー
- * @class repository.Stock
+ * 在庫数カウント結果インターフェース
+ */
+export interface ICountResult {
+    _id: string;
+    count: number;
+}
+
+/**
+ * 在庫リポジトリ
  */
 export class MongoRepository {
-    public readonly stockModel: typeof StockModel;
+    private readonly stockModel: typeof StockModel;
 
     constructor(connection: Connection) {
         this.stockModel = connection.model(StockModel.modelName);
@@ -16,7 +23,6 @@ export class MongoRepository {
 
     /**
      * まだなければ保管する
-     * @param {factory.stock.IStock} stock
      */
     public async saveIfNotExists(stock: factory.stock.IStock) {
         await this.stockModel.findOneAndUpdate(
@@ -32,6 +38,70 @@ export class MongoRepository {
             {
                 upsert: true,
                 new: true
+            }
+        ).exec();
+    }
+
+    /**
+     * 在庫数をカウントする
+     */
+    public async count(params: {
+        availability: factory.itemAvailability;
+        performance: { ids: string[] };
+    }): Promise<ICountResult[]> {
+        return this.stockModel.aggregate(
+            [
+                {
+                    $match: {
+                        availability: params.availability,
+                        performance: { $in: params.performance.ids }
+                    }
+                },
+                {
+                    $group: {
+                        _id: '$performance',
+                        count: { $sum: 1 }
+                    }
+                }
+            ]
+        ).exec();
+    }
+
+    /**
+     * 在庫おさえ
+     */
+    public async lock(params: {
+        performance: string;
+        holder: string;
+    }): Promise<factory.stock.IStock | null> {
+        const doc = await this.stockModel.findOneAndUpdate(
+            {
+                performance: params.performance,
+                availability: factory.itemAvailability.InStock
+            },
+            {
+                availability: factory.itemAvailability.OutOfStock,
+                holder: params.holder
+            },
+            { new: true }
+        ).exec();
+
+        return (doc !== null) ? doc.toObject() : null;
+    }
+
+    /**
+     * 在庫解放
+     */
+    public async unlock(params: factory.reservation.event.IStock) {
+        await this.stockModel.findOneAndUpdate(
+            {
+                _id: params.id,
+                availability: params.availability_after,
+                holder: params.holder // 対象取引に保持されている
+            },
+            {
+                $set: { availability: params.availability_before },
+                $unset: { holder: 1 }
             }
         ).exec();
     }