+++ CHANGELOG.md
@@ -12,7 +12,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 - パフォーマンス検索条件から作品関係を削除
 - パフォーマンスコレクションからリレーションを削除
-- 予約リポジトリのmongooseモデルをprivate化
+- 在庫リポジトリのmongooseモデルをprivate化
 
 ### Deprecated